processor/requirements.txt
influxdb-client==1.45.0
paho-mqtt==2.1.0
python-dotenv==1.0.1